refactor: contents ui 스트림릿 리펙토링
