+++ my_app/ui/contents/__init__.py
@@ -1,0 +1,1 @@
+# UI Contents Package
